HYD-5937 GUI: Resettable group $addControl being called on incorrect form
When monkey patching the $addControl function on a form control the
oldAddControl associated with the parent form is being used to overwrite
the $addControl function on a nested form. Instead, the oldAddControl
function should be the function on the nested form. Without doing this,
the controls in the nested forms will not be added to the nested form
controller.

This patch will update the resettable-group component such that when
overwriting the $addControl function on a form controller, the original
$addControl function will be saved off from each form and will then be
called respectively when that form's addControl function is called.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>