HYD-5949 GUI: Abstract configToData$ logic to be used as a consumer between all charts
Since updating the charts to add persistence and tie them into the
store, I've noticed that each chart needs a config data stream to chart
data stream consumer. The logic is very similar when creating this
consumer in each of the charts and thus can be abstracted.

In this patch a new file will be created to handle data config to chart
data stream consumption.

- Added config-to-data.js factory
- Added config-to-data-test.js to test the consumer

Signed-off-by: Will Johnson <william.c.johnson@intel.com>