HYD-5974 GUI: Add "less@2.6.1" dependency to package.json
GUI is currently failing because gulp-less is failing during the build
process. This is because gulp-less is using less 2.7.0, which currently
has a bug. To get around the build issue, we can add less as a
dependency and set it to 2.6.1.

This patch will update package.json such that less@2.6.1 is added as a
dependency.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>