HYD-6015 Autocomplete error "a is not a function" on status page
The tokens parser was trying to use a non-existant method
in an inList call.

This was causing the parser to throw an error whenever a string
containing:

type in [

was typed.

Switch the object to the one that has the method we are looking for.
We have also added a missing @flow pragma to tokens.js which causes this
error to not typecheck.

  - Update to use the correct object.

Signed-off-by: Joe Grund <joe.grund@intel.com>