HYD-6058 GUI: Add known types from chart-persistence branch to master branch
The chart persistence branch has introduced numerous types that would be
useful for master and other development branches. It would be nice to
merge these types into the master branch so they can be used on other
features.

This patch will introduce known types from the persistent charts branch
to the master branch.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>