HYD-6098 GUI: Add chart transformers module
While adding persistence to the dashboard charts we discovered a pattern
that can be abstracted into a chart transformer. It would make sense to
create a chart transformer module that can be used between charts to
transform data as it streams through.

This patch will create the chart transforms module and update types
throughout other modules as necessary.

- Added chart-transformers module and chart transformer
- Updated types in modules that will be used for charting
- Unit tests still need to be updated/added to this patch