HYD-6186 Upgrade Highland to 3.x
Update Highland to 3.x.

This entails removing properties and instead using a broadcaster.

This is because properties messed with internal stream state which no
longer exists.

By using a broadcaster we can implement a simple pub / sub event-stream
exchange.

This is a significant change as properties are pervasive throughout the
codebase.

We can also remove rebind destroy, as highland 3.x does this implicitly.

Signed-off-by: Joe Grund <joe.grund@intel.com>