HYD-6156: Add Breadcrumb component
One of the 3.1 features is to add breadcrumbs to each of the pages in
the application. The breadcrumbs should keep track of the users history
as they navigate through the app and display a linke to each of the
previous pages. When selecting a link on a page, a new item will be
added to the breadcrumb while going back should pop an item off the
breadcrumb component. Cycles will be eliminated by slicing up to the
point of the existing breadcrumb node. Additionally, the breadcrumbs
will be reset when clicking on menu or tree items.

This patch will:
1. Create the breadcrumb component
2. Add the breadcrumb component to each of the pages
3. Update links to indicate where the breadcrumb stack should be reset
4. Create the page header component and remove existing page headers

Signed-off-by: Will Johnson <william.c.johnson@intel.com>