HYD-6204 Multiple scroll bars in GUI with tree-view in iframe
Some of the GUI now lives in an iframe.

An iframe does not conform to it's parent height automatically; it has
it's own height.

Because of this, we see multiple scrollbars when the iframe is larger
than it's containing viewport.

Use the following statement in a setInterval:
frame.style.height = frame.contentWindow.document.body.scrollHeight +
'px';

This will sync up the frame height with it's container, and we will see
a single scrollbar.

Signed-off-by: Joe Grund <joe.grund@intel.com>