HYD-6156 Create Promise Transforms
The breadcrumb component will utilize resolves that contain data related
to the route states. To obtain the resolved data, a couple of promise
transforms can be utilized:

1. streamToPromise - converts a stream to a promise
2. extendParentData - adds parent data to a promises data

In addition to creating promise-transforms, resolveStream can be moved
into the promise-transforms module. This means that all modules using
resolveStream will need to be updated.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>