HYD-6257 GUI: add reset param to menu / tree-view links
There are some links in the GUI that should not push another breadcrumb
on the stack during transition.

Namely, these are the menu and tree-view links, which can be clicked out
of context of the current page.

Update all relevant links to use ui-sref with a resetState param set to
true.

  - Update all necessary menu, footer
    and tree-view links.

Signed-off-by: Joe Grund <joe.grund@intel.com>