HYD-6256 Add top-like component for jobstats
Create a top-like component for viewing jobstats.

It should be a "realtime" table where top users are easily displayed.

The ideal use case is for an operator to view the table and quickly see
any jobs causing problems.

Architecture
Currently jobstats are fetchable one metric at a time. This means to
display an aggregate, we need to:

1) Do four fetches for each metric type (read bytes, write bytes, read
IOPS, write IOPS). These fetches should happen every 10s or so.

2) Average all the points across all OSTs grouped by id.

3) Reduce the four data-sets into single rows.

4) Determine a suitable sorting strategy to show top users.

Probably will be useful to query the list as well, but that may need
more thought.

Signed-off-by: Joe Grund <joe.grund@intel.com>