HYD-6279 Add animations between state transitions
When changing between states the views pop up instantly. This isn't a
bad thing, but adding a small, subtle animation that renders the state
will allow the user to transition between pages with a better user
experience.

This patch will add a ui-loader-view directive that will add a ui-view
in its template but handle fading in and out state change animations.
This is especially useful when resolves are involved in the changing
state. It also provides the ability to ensure that a spinner is only
loaded one time and that a spinner may or may not show based on if that
route change is the same. This is useful when there is a subview and you
don't want to show the spinner for the app view.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>