Log page is picking up qs from previous route
The log page is currently picking up qs from a previous route, causing
an error. This is due to qs-from-location using the $location service to
get the query string off the url. The problem is that the $location
service will get the url in the "from" part of the transition.

This patch will update qs-from-location to grab the query string from
the "to" part of the route. It will also update qs-stream such that the
current state parameters are being passed in correctly. To get the
current state parameters, $stateParams must be defined on either a
resolve or a controller. We can also obtain the current parameter hash
from a Transition object directly while in a transition hook.

Fixes #477.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>