HYD-6388 Provide anonymous jobstats to users that are not logged in
Jobstats should be available to users that are not logged in. When a
user accesses the app as an anonymous user they should be able to click
on the Jobstats menu item and view the jobstats duration picker as well
as the jobstats table.

This patch will remove the access restriction from the jobstats menu
item as well as the access restriction in the jobstats state.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>