GUI-#516 Set page overflow-y to be hidden only in desktop mode
The overflow-y: hidden style is defined for the entire app regardless of
the view. We only want this to be hidden in the desktop view, otherwise,
mobile devices will not be able to scroll.

This patch updates the main.less file such that the overflow-y property
is set to hidden in desktop mode only.

Fixes #516.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>