GUI-#514 Clicking the begin link on the status page throws an error
Since upgrading to UI-Router, query parameters now have more encoding.
In particular, comma's and colon's have been encoded to %2C and %3A
respectively. The qs parser, however, is expecting the comma and colon
to NOT be encoded, thus causing an error in cases like this:

https://localhost:8000/ui/log/?datetime__gte=2016-08-29%2015%3A44%3A08
and
https://localhost:8000/ui/status/?severity__in=WARNING%2CERROR&active=true

In these two cases, an error would be thrown because the parser doesn't
understand what %2C and %3A are in the links above. The parser is now
updated to handle these strings in intel-qs-parsers@3.1.2. Therefore,
we should no longer call decodeURIComponent on the query parameters in
qs-from-location.js as we want the string to remain encoded. All related
tests should also be updated such that commas and colons are converted
to %2C and %3A respectively.

Fixes #514.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>