GUI-#515 Fix page title and breadcrumb popping during page transition
The breadcrumb and page title disappear for a second while the page
transitions. This looks very strange and should instead blend in with
the page transition.

This patch will make it so that the page title and breadcrumb text fade
to an opacity of .2 during the transition and then fade back once the
transition completes.

Fixes #515.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>