GUI-#529 Fix race condition on users page
Since the user data is not cached it is being fetched over the wire. If
the data does not come back before attempting to retrieve it from the
store then it will be undefined. If undefined, the user's page will not
load the users appropriately. For now, we should filter on the stream
returned by the store to make sure that it has data before we proceed
down the transform.

This patch updates the old-gui-resolve to ensure that the data returned
by the user store is not empty. If it is empty, it will end the
transforms and prevent an error from being thrown.

Fixes #529.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>