GUI-#541 Create confirm button component
Rather than display a modal to confirm an action, it is more user
friendly to simply change the state of a button from something such as
delete to confirm delete. This has the benefit of forcing the user to
confirm their action without having to create an extra interface that is
distracting to the user. This component will used multi slot
transclusion, allowing the developer to add the two button states, which
will be unique to the feature they are working on. This makes the
component more flexible.

This patch will create the confirm button component and add it to the
iml module.

Fixes #541.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>