HYD-6477 Out of sync between the file system tab and the detail file system view
Please see the attached screenshots. When one MDS is down, the detail
file system view showed that the MGT is not running on the 2nd MDS,
kmds02. However, when I clicked the file system link on the left side
bar, the simplified file system view still shows kmds01 as the active
Management server and also the status with "No issues".

This patch will update the table header from "Management Server" to "Primary
Management Server" on the filesystems page.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>