HYD-6502 Expand minimum jobstats range to 30 seconds
When you click on the heatmap it brings up the jobstats page with a
range of 16 seconds. This is too small and results in the page showing a
message that jobstats are unconfigured.

Manually expanding the range shows jobstats on the page.

The default range was expanded in a previous ticket.

Make the default range large enough that the page doesn't show
the unconfigured message to avoid user confusion.

  - Expand range to 30 second minimum.

Signed-off-by: Joe Grund <joe.grund@intel.com>