HYD-6526 Update gulp to not use symlinks
While running the IML manager in a Docker container via a mount, symlinks are not picked up correctly.

This means files appear to be empty in Docker and we get 404s when trying to load pages during dev.

Instead, change the symlinks to file writes, which Docker sees correctly.

In the longer term, we can probably just bind the GUI dir into Docker directly and not worry about copying the files.

  - Change all usages of gulp.symlink to gulp.dest.

Signed-off-by: Joe Grund <joe.grund@intel.com>