HYD-6855 Create connect-to-store component
Create a higher order component that allows a component to connect to
the store directly.

The syntax will look like:
connectToStore('loginForm', ({loginForm}) =>
  <SomeEl someData={loginForm} />
  );

Where `loginForm` is the props coming from the store.

Behind the scenes this component will use select to sync with the
specified key.

Whenever the data changes, the component will re-render.
When the component is unmounted, the select will be destroyed.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>