HYD-6926 GUI: Closing modal when adding a server produces an error
While testing the weekly build I received an error when closing one of
the modals in the three step process of adding a server. This patch will
address and resolve the error.

This patch updates wait-for-command-completion-service.js such that the
commandModal$ is created from the command$ fork. This forked stream is
then destroyed. A test has been added to ensure that the forked stream
is destroy and that the original stream where it was forked from is not
destroyed.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>