HYD-6933 GUI: Create angular-exec
We have run into an issue in multiple cases where we need to use the router to go to a new location in the app but because we are importing dependencies rather than using angular's DI we do not have access to the ui-router state service, which allows us to navigate to a new state of the app. This ticket will address this issue by creating a new module, angular-exec, which will allow you to specify the service you need to use, the method to execute on the service, and its associated args. The result of executing the method on the service will be returned in a stream.

angularExec('<serviceName>', '<method>', ['param1', ['param2', ...]]) => HighlandStreamT<R>
angularExec('$state', 'go', 'app.dashboard.overview');

Signed-off-by: Will Johnson <william.c.johnson@intel.com>