Add route for storage plugin screen
The angularjs route for the IML configuration -> storage page is not consistent with the method used for the
other routes used in the gui repository. The gui repository uses the method used by the angular-ui-routes
library.

Create the necessary module, states, controllers, etc. to create a new storage route.

 - Added the basic route functionality for storage
 - Created a storage-module, storage-component and storage-states
 - Created basic jasmine test cases for storage-component and storage-states
 - Removed code that implemented the storage route in the old way
 - Passed eslint and test

Signed-off-by: Peter Robinson <peter.timothyx.robinson@intel.com>

Conflicts:
	source/iml/iml-module.js
	source/iml/old-gui-shim/old-gui-resolves.js
	source/iml/old-gui-shim/old-gui-states.js