Remove all inline usages of angular reset and jest mock reset.
We are currently preserving state between test runs. We do not want to
do this; it will make our tests quite hard to debug.

Instead, make sure we are reseting modules and mocks between each test.
This may be slightly slower, but will give us assurances we are not leaking state.

The reason why resetting mocks was removed is because ui-bootstrap is expecting angular to be present after
we remove it between tests.

To get around it, we have three options:

1. We don’t use angular anywhere in source besides the main registration module.
2. We mock modules and reset angular. Anywhere we need angular later on,
we make sure we load it in a beforeEach.
3. We don’t mock modules and angular. We do it explicitly in tests when needed

We were currently taking option 3, this patch moves us to option 2.

The ultimate goal is to do option 1. We are pretty close to this,
we just need a merge fn and to remove a few more modules.

That fix can be later on.

Signed-off-by: Joe Grund <joe.grund@intel.com>