Fix missing unwrap of maybe in legend (#58)
* Bump deps

Bump deps to latest.

In addition, fix heat map test that was broken by nvd3 bump.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Fix missing unwrap of maybe in legend

The leged was not unwrapping a maybe type
being returned by `fp.last`. Due to this,
it would always hide the labels in the
legend. This was being missed because
there was no flow pragma on this file.

Add the flow pragma and use `maybe.matchWith`
to unwrap the maybe.

In addition add some more d3 types, and fix
spacing in the legend tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>