Emit resize event on panel size change (#59)
* Bump deps

Bump deps to latest.

In addition, fix heat map test that was broken by nvd3 bump.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Fix missing unwrap of maybe in legend

The leged was not unwrapping a maybe type
being returned by `fp.last`. Due to this,
it would always hide the labels in the
legend. This was being missed because
there was no flow pragma on this file.

Add the flow pragma and use `maybe.matchWith`
to unwrap the maybe.

In addition add some more d3 types, and fix
spacing in the legend tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Emit resize event on panel size change

Now that we are using both angular, and
moving to inferno, we should
use native event dispatch
to bridge the gap between
the two frameworks. use window.dispatchEvent
when the panel is being resized,
so we can pick up on it natively.

Signed-off-by: Joe Grund <joe.grund@intel.com>