Implement Chart components for inferno. (#63)
Implement some chart components for inferno.

There are a few rules that make this easier.

1. Only pass changes downward, no handlers for
chart updating on the chart component itself
2. We skip scales as a component. They are
the variant part of making a chart in many
cases, we can just create them inline during
usage.
3. Each component will pass it's props, and useful
parent props to their children. This ensures that
scales and such are propagated where they should be.