GH-12 Fix dispatch source usage with ALLOW_ANONYMOUS_READ
We currently gate dispatch source streams with ALLOW_ANONYMOUS_READ.

This is insufficient in the case that ALLOW_ANONYMOUS_READ is false.

In that case, the stream will never be called, regardless or whether the
user has logged in or not.

Instead there should be an or check of ALLOW_ANONYMOUS_READ ||
CACHE_INITIAL_DATA.session.user

Signed-off-by: Will Johnson <william.c.johnson@intel.com>