Log States stream hangs on exit
When leaving the log page both the socket stream and the store stream
hang and do not end. This is because we are using zip to combine the
servers stream (from the store) and the socket stream (/log). zip is
designed to work with finite streams and does not end the streams as we
expected. Instead, we should use "multi-stream", which will destroy the
streams after leaving the log page.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>