Update Formatting (#122)
Description:
The prettier plugin in VSCode does not match the version specified in
package.json. Eslint and its dependencies should be updatedto their
latest version. This will allow formatting in vscode to match what
eslint is expecting.

In addition to formatting, eslintignore is currently ignoring the "dist"
directory. It should instead be ignoring the `targetdir` directory as
well as the `include` directory.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>