Turn autoFreeze off in production
While immer detects the node environment our web application is not
running with Node. It is, however, run through webpack which does take
the NODE_ENV environment variable into account. In the current state,
however, webpack does not transform environment checks during the
transpilation phase. To resolve this, i've added the webpack
EnvironmentPlugin, which will take the following:

const isProduction = process.env.NODE_ENV === 'production'

and substitute the environment value when running webpack. For example,
say we build with the following:

```
NODE_ENV=production webpack -p
```

The code would be transpiled as follows:

```
const isProduction = 'production' === 'production'
```

This can then be used to determine whether or not to set autoFreeze on
immer.

Note that immer also has a built in check that is similar:

```
var inProduction =
  (typeof process !== "undefined" && process.env.NODE_ENV === "production") || verifyMinified.name !== "verifyMinified";
```

After some testing I discovered that this doesn't give us the result we
want. In my case, I tested the "development" environment but running
webpack with NODE_ENV=development. While process.env.NODE_ENV came out
to be "development", `inProduction` was still set to true, most likely
due to the OR statement. I decided against relying on this and manually
set autoFreeze in iml-module.js depending on the value of NODE_ENV
during transpilation.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>