Fix Backoff for SSE (#192)
Testing backoff for server send events currently works when it receives 502's. However, if nginx is stopped or the
chroma manager is restarted, the retry count will skyrocket exponentially. This was due to the event stream being closed
inside the setTimeout. To fix this, the close was moved outside of the timeout and only "next" is called inside the
timeout to turn the generator.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>