Available updates button disabled (#228)
* Available updates button disabled

Fixes #1237.

Description:
The available updates button remains disabled in an upgrade to 5.1. This is due to warp-drive not picking up changes for the computed property member_of_active_filesystem (which is derived mostly from the fs table). It seems like the most straightforward way to solve this is to check for server fs membership in the GUI by joining the data sets together.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* - Cleanups
- Add testing

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* - remove unnecessary combinator

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>