GH-232 Selecting filesystem from dashboard dropdown causes error
Fixes #232.

When selecting the filter on the dashboard page and selecting a filesystem the gui will crash. This happens because the
data format has changed from an array of objects to a hash in which each key points to an object. The data is
essentially the same, but in a different format. This caused the filter to blow up when selecting a filesystem,
            or applying a target. In addition, once the error was resolved, all filesystems were being displayed even
            though an individual one was selected. This happened because the dashboard controller was not filtering out
            filesystems based on the filesystem id that was selected. This patch resolves each of these problems.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>